requirements.txt
asgiref
attrs
backoff
blinker
branca
Brotli
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
dj-database-url
dj-rest-auth
Django
django-crispy-forms
django-environ
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
folium
h11
h2
hpack
hyperframe
idna
Jinja2
kaitaistruct
Markdown
MarkupSafe
MouseInfo
mysql-connector
mysql-connector-python
numpy
oauthlib
opencage
outcome
packaging
phonenumbers
pillow
psycopg2
psycopg2-binary
pyasn1
PyAutoGUI
pycparser
pydivert
PyGetWindow
PyJWT
PyMsgBox
pyOpenSSL
pyparsing
pyperclip
PyRect
PyScreeze
PySocks
python-dotenv
python3-openid
pytweening
requests
requests-oauthlib
selenium-wire
sniffio
social-auth-app-django
social-auth-core
sortedcontainers
sqlparse
trio
trio-websocket
typing_extensions
tzdata
tzdata
urllib3
webdriver-manager
wsproto
xyzservices
zstandard